feat: 租户人员完善

+++ src/views/sys/sso-tenant/index.vue
@@ -186,9 +186,11 @@
        * @param
        */
       function handleDelete(ssoTenant: SsoTenant) {
-        deleteSsoTenant(ssoTenant.id).then(() => {
-          handleSuccess();
-        });
+        if (ssoTenant.id) {
+          deleteSsoTenant(ssoTenant.id).then(() => {
+            handleSuccess();
+          });
+        }
       }
 
       /**
